feat: general UI refactor as well as QoL upgrade
Improved the vertical scroll available on the actual website
so that there is not unecessary scroll pages.

Standardized the usage of plain grey instead of white colour on
the page.

Removed many nav items and bars to open up the space used for the
tool directly.

Added smoother scrolling on the multiple user input for the tool